Add table "FestivalPerformer". Add trigger file inside sql folder.
Trigger functions creates new column in "FestivalPerformer" which is type of DATERANGE. It creates new column beacuse there is exclude constraint which uses general tree to serach if there is row with same festival period and performer id.
Performer can not participate in festivals which are in the same period.